+++ src/Crypt.php
@@ -274,8 +274,8 @@ class Crypt {
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
-     * @noinspection PhpUnused
-     */
+	 * @noinspection PhpUnused
+	 */
 	public function set_plugin_file( string $plugin_file ): void {
 		$this->plugin_file = $plugin_file;
 	}
@@ -582,8 +582,8 @@ class Crypt {
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
-     * @noinspection PhpUnused
-     */
+	 * @noinspection PhpUnused
+	 */
 	public function clear_errors(): void {
 		$this->errors = null;
 	}